fix: before_first_request deprecation
