+++ setup.cfg
@@ -89,6 +89,9 @@ invenio_users_resources.moderation.actions =
     block = invenio_communities.requests.user_moderation.actions:on_block
     restore = invenio_communities.requests.user_moderation.actions:on_restore
     approve = invenio_communities.requests.user_moderation.actions:on_approve
+invenio_base.finalize_app =
+    invenio_communities = invenio_communities.finalize_app:finalize_app
+
 
 [build_sphinx]
 source-dir = docs/
